Adjustment of tables and graphs
Table A1 was adjusted, modifying its construction and style. Also, color was added to the graphs and the option to highlight them in each execution was added. PCA graphs were saved manually, because the library does not allow it. Also, an informative paragraph was added about the link of the table with the conference poster.